% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{add_organelle_proteins}
\alias{add_organelle_proteins}
\title{Add organelle ID lists for each proteome}
\usage{
add_organelle_proteins(strata, overwrite = FALSE)
}
\arguments{
\item{strata}{A Strata object where leafs are NCBI taxa represented in UniProt}

\item{overwrite}{logical. If TRUE, then the 'proteome_stats' field will be
overwritten if it already exists.}
}
\value{
A Strata object with a new 'organelle' data field
}
\description{
Add organelle ID lists for each proteome
}
