% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{add_representative}
\alias{add_representative}
\title{Add representatives to the strata}
\usage{
add_representative(cousin_sets, scinames = NULL, taxids = NULL)
}
\arguments{
\item{cousin_sets}{A nested list of cousins}

\item{scinames}{scientific names of representatives to add}

\item{taxids}{NCBI ids of representatives to add}
}
\description{
Add representatives to the strata
}
