% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{add_taxid}
\alias{add_taxid}
\title{Add an id to a representative list}
\usage{
add_taxid(strata, taxid)
}
\arguments{
\item{strata}{A Strata object}

\item{taxid}{NCBI taxon ID of the species to add}
}
\value{
Strata object
}
\description{
Add an id to a representative list
}
