% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{add_to}
\alias{add_to}
\title{Add an id to a representative list}
\usage{
add_to(cousin_sets, stratum_id, uncle_id, new_id)
}
\arguments{
\item{cousin_sets}{A nested list of cousins}

\item{stratum_id}{Stratum NCBI taxonomy id}

\item{uncle_id}{Uncle NCBI taxonomy id}

\item{new_id}{The ID to be added}
}
\description{
Add an id to a representative list
}
