% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic_functions.R
\name{ancestors}
\alias{ancestors}
\title{Get the lineage of a given species}
\usage{
ancestors(taxid, byname = FALSE)
}
\arguments{
\item{taxid}{a single NCBI taxon}

\item{byname}{Get ancestor names, instead of NCBI ids}
}
\description{
Get the lineage of a given species
}
