% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arabidopsis_strata}
\alias{arabidopsis_strata}
\title{An Arabidopsis thaliana based Strata object}
\format{Strata}
\usage{
arabidopsis_strata
}
\description{
132 species are included in the tree. No data has been loaded.
}
\keyword{datasets}
