% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{as_named_strata}
\alias{as_named_strata}
\title{Substitute names for taxids in strata}
\usage{
as_named_strata(x, scinames = NULL)
}
\arguments{
\item{x}{data.tree, where node names are NCBI taxon IDs}

\item{scinames}{named character vector, where names are NCBI taxon IDs and
elements are scientific names}
}
\value{
data.tree, where all names are scientific names
}
\description{
Be careful with this function; data.tree uses reference semantics so this
function can have unintended side-effects.
}
