% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atree}
\alias{atree}
\title{A simple, bifurcating tree for use in util examples}
\format{phylo}
\usage{
atree
}
\description{
A simple, bifurcating tree for use in util examples
}
\examples{
# Here is the code used to generate this tree
set.seed(42)
atree <- rtree(10)
atree$tip.label <- paste0("t", 1:ntip(atree))
atree$node.label <- paste0("n", 1:a$Nnode + ntip(atree))
}
\keyword{datasets}
