% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{classify_assuming_iid}
\alias{classify_assuming_iid}
\title{Classify homologs under the independence assumption}
\usage{
classify_assuming_iid(threshold, ...)
}
\arguments{
\item{threshold}{The target alpha (e.g. 0.05)}

\item{...}{Additional arguments passed to p.adjust}
}
\description{
Classify homologs under the independence assumption
}
