% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{classify_by_adjusted_pvalue}
\alias{classify_by_adjusted_pvalue}
\title{Infer homology inference based on an adjusted p-value threshold}
\usage{
classify_by_adjusted_pvalue(threshold, ...)
}
\arguments{
\item{threshold}{An e-value threshold (before adjustment)}

\item{...}{Unused}
}
\value{
A function of a dataframe with columns 'evalue' and 'staxid'
}
\description{
Here the E-values of sequence similarity against each target species is
multiplied by the total number of target species. This is equivalent to what
BLAST does in the special case where all proteomes are of equal length.
}
