% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{classify_by_evalue}
\alias{classify_by_evalue}
\title{Infer homology inference based on a hard e-value threshold}
\usage{
classify_by_evalue(threshold)
}
\arguments{
\item{threshold}{An e-value threshold}
}
\value{
A function of a dataframe that has the column 'evalue', the function
}
\description{
Infer homology inference based on a hard e-value threshold
}
