% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{classify_by_fisher}
\alias{classify_by_fisher}
\title{Infer homology based on p-value with Fisher's method}
\usage{
classify_by_fisher(threshold)
}
\arguments{
\item{threshold}{P-value threshold}
}
\description{
Infer homology based on p-value with Fisher's method
}
