% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{classify_by_pvalue}
\alias{classify_by_pvalue}
\title{Infer homology inference based on a hard p-value threshold}
\usage{
classify_by_pvalue(threshold, ...)
}
\arguments{
\item{threshold}{An e-value threshold}

\item{...}{Unused}
}
\value{
A function of a dataframe that has the column 'evalue', the function
}
\description{
Infer homology inference based on a hard p-value threshold
}
