% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{clear_aunt}
\alias{clear_aunt}
\title{Clear all representatives of a given aunt}
\usage{
clear_aunt(cousin_sets, stratum_id, aunt_id)
}
\arguments{
\item{cousin_sets}{A nested list of cousins}

\item{stratum_id}{Stratum NCBI taxonomy id}

\item{aunt_id}{aunt NCBI taxonomy id}
}
\description{
Clear all representatives of a given aunt
}
