% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{clear_stratum}
\alias{clear_stratum}
\title{Clear all representatives of a given stratum}
\usage{
clear_stratum(cousin_sets, stratum_id)
}
\arguments{
\item{cousin_sets}{A nested list of cousins}

\item{stratum_id}{Stratum NCBI taxonomy id}
}
\description{
Clear all representatives of a given stratum
}
