% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{clear_uncle}
\alias{clear_uncle}
\title{Clear all representatives of a given uncle}
\usage{
clear_uncle(cousin_sets, stratum_id, uncle_id)
}
\arguments{
\item{cousin_sets}{A nested list of cousins}

\item{stratum_id}{Stratum NCBI taxonomy id}

\item{uncle_id}{Uncle NCBI taxonomy id}
}
\description{
Clear all representatives of a given uncle
}
