% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{constrict}
\alias{constrict}
\title{Find the focal sequences that are not represented in the given strata}
\usage{
constrict(d, on, revenants = NULL, ...)
}
\arguments{
\item{d}{data.frame of maximal hits}

\item{on}{The strata levels (NOT taxon IDs) on which to constrict}

\item{revenants}{A table of the form created by \code{find_revenants}. If
NULL, a new table is created.}

\item{...}{Additional arguments sent to \code{find_revenants} if a new
revenant table is being created (i.e. 'revenants=NULL').}
}
\value{
A subset of d where no gene has a match to the strata specified in 'on'
}
\description{
Find the focal sequences that are not represented in the given strata
}
