% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{CountMatrix}
\alias{CountMatrix}
\title{Initialize a CountMatrix object}
\usage{
CountMatrix(x, ylab = "from", xlab = "to")
}
\arguments{
\item{x}{The matrix}

\item{ylab}{A label for the y axis (used in plotting, for example)}

\item{xlab}{A label for the x axis}
}
\value{
CountMatrix object
}
\description{
Initialize a CountMatrix object
}
