% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic_functions.R
\name{cousins}
\alias{cousins}
\title{Get the species level cousins of a taxid}
\usage{
cousins(taxid)
}
\arguments{
\item{taxid}{a single NCBI taxon}
}
\description{
Get the species level cousins of a taxid
}
