% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylostratr.R
\name{data_tree_to_result_tree}
\alias{data_tree_to_result_tree}
\title{Convert sequence tree to result tree}
\usage{
data_tree_to_result_tree(x, FUN, ...)
}
\arguments{
\item{x}{Tree of protein FASTA sequences}

\item{FUN}{A function of the focal species raw data and a single subject
species raw data that produces a raw comparison result}

\item{...}{Additional arguments passed to FUN}
}
\value{
A tree of raw results
}
\description{
Convert sequence tree to result tree
}
