% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{descendents}
\alias{descendents}
\alias{descendents.Strata}
\alias{descendents.phylo}
\title{Get all leafs descending from a set of ids}
\usage{
descendents(x, ...)

\method{descendents}{Strata}(x, ...)

\method{descendents}{phylo}(x, id, ...)
}
\arguments{
\item{x}{phylo object}

\item{...}{Arguments sent to \code{descendent_nodes}}

\item{id}{vector of ids or names}
}
\value{
integer vector of leaf ids
}
\description{
This is a simple wrapper around \code{descendent_nodes}.
}
