% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{diverse_subtree}
\alias{diverse_subtree}
\title{Get a diverse subset of species from tree}
\usage{
diverse_subtree(tree, n, weights = NULL, collapse = FALSE, FUN = .algo1,
  ...)
}
\arguments{
\item{tree}{phylo object}

\item{n}{number of species to keep (if greater then the number of species in
the tree, the tree is returned unchanged)}

\item{weights}{Numeric vector with length equal to the number of species in
the tree. A weight of 1 will have no influence, lower than one means the
species is less likely to be selected.}

\item{collapse}{Should nodes with a single descendent be collapse?}

\item{FUN}{The diversification algorithm to use}

\item{...}{Additional arguments passed to the diversification algorithm function}
}
\value{
phylo object
}
\description{
Get a diverse subset of species from tree
}
\examples{
data(atree)

# default
diverse_subtree(atree, 4)

# do not to include the 't1' species
diverse_subtree(atree, 4, weights=c(t1=0))
}
