% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{do_on}
\alias{do_on}
\title{Map a function over a specific stratum}
\usage{
do_on(cousin_sets, stratum_id, fun)
}
\arguments{
\item{cousin_sets}{A nested list of cousins}

\item{stratum_id}{Stratum NCBI taxonomy id}

\item{fun}{A function of the uncle list}
}
\description{
Map a function over a specific stratum
}
