% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edgelist_to_phylo}
\alias{edgelist_to_phylo}
\title{Make a phylo object from an edgelist}
\usage{
edgelist_to_phylo(edgelist)
}
\arguments{
\item{edgelist}{A matrix of node names}
}
\value{
phylo object
}
\description{
Make a phylo object from an edgelist
}
\examples{
edgelist <- matrix(c('A', 'A', 'B', 'B', 'B', 'C', 'D', 'E', 'A', 'F'), ncol=2)
tree <- edgelist_to_phylo(edgelist)
# A bug in ape causes a segfault here, but the French bastards in charge don't
# seem to take the issues seriously
# plot(tree, show.node.label=TRUE)
}
