% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{extract_uniprot_id_from_fasta}
\alias{extract_uniprot_id_from_fasta}
\title{Parse a UniProt ID from a FASTA file}
\usage{
extract_uniprot_id_from_fasta(fastafile)
}
\arguments{
\item{fastafile}{A protein FASTA file, assumed to be UniProt format}
}
\value{
character vector of UniProt IDs IF this is a UniProt FASTA, NULL otherwise
}
\description{
UniProt FASTA files have the initial pattern:
  <db>|<UniProt ID>|<alternative ID> <desc>'
}
\details{
Here we parse out the UniProt ID. If the FASTA header does not match the
pattern, we return NULL
}
