% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{find_revenants}
\alias{find_revenants}
\title{Find genes that with matches that skip strata}
\usage{
find_revenants(d, classifier = classify_by_evalue(1e-05))
}
\arguments{
\item{d}{data.frame of maximal hits}

\item{classifier}{function of a blast result that infers whether each
subject species contains a homolog}
}
\value{
data.frame with columns qseqid and ps, where ps represents the
strata that are skipped
}
\description{
Find genes that with matches that skip strata
}
