% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prokaryotes.R
\name{generate_prokaryote_tree}
\alias{generate_prokaryote_tree}
\title{Generate a custom prokaryote phylogenetic tree}
\usage{
generate_prokaryote_tree(
  domain.n = 50,
  weights = NULL,
  focal_taxid = "3702",
  current.domain,
  drop.names = NULL
)
}
\arguments{
\item{domain.n}{approximate number of species desired from this domain (may not give exactly this tree size due to rounding)}

\item{weights}{optional named numeric vector of weights to use for sampling, names are Uniprot taxon IDs}

\item{focal_taxid}{focal taxa just needed to convert to strata object; not used otherwise}

\item{current.domain}{choose domain id from UniProt: '2' = bacteria, '2157' = archaea}

\item{drop.names}{optional character vector of taxa names (Uniprot taxon IDs) to drop entirely}
}
\value{
phylo object containing the prokaryotic tree
}
\description{
Creates a diverse prokaryotic phylogenetic tree of desired size with optional weighting and the ability to exclude specified taxa.
Based on diverse_subtree but applied across the whole tree rather than based on a focal taxon.
}
