% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{get_besthit}
\alias{get_besthit}
\title{Get the "besthit" table for a given species}
\usage{
get_besthit(strata, taxid)
}
\arguments{
\item{strata}{Strata object}

\item{taxid}{A species name or taxonomy ID}
}
\value{
data.frame of best hits (\code{get_max_hit})
}
\description{
Get the "besthit" table for a given species
}
\examples{
\dontrun{
strata@data$besthits[["Unicorn"]] <- get_besthit(strata, "Unicorn")
}
}
