% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{get_mrca_names}
\alias{get_mrca_names}
\title{Get an ordered factor mapping MRCA taxon IDs (as vector names) to names}
\usage{
get_mrca_names(hittable)
}
\arguments{
\item{hittable}{A data.frame with the columns 'ps' and 'mrca'}
}
\description{
Get an ordered factor mapping MRCA taxon IDs (as vector names) to names
}
