% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{get_phylostrata_map}
\alias{get_phylostrata_map}
\title{Get a phylostrata map from a Strata object}
\usage{
get_phylostrata_map(strata)
}
\arguments{
\item{strata}{Strata object}
}
\value{
data.frame with columns 'species', 'mrca', an 'ps', where 'mrca' is
a factor of phylostrata with levels ordered by age
}
\description{
Get a phylostrata map from a Strata object
}
