% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{get_root}
\alias{get_root}
\title{Get the root of a tree}
\usage{
get_root(tree)
}
\arguments{
\item{tree}{phylo object}
}
\value{
index of the root node (or nodes if there are more than one)
}
\description{
There should normally be only one, but if there are more, this function will
find all of them.
}
