% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{has_header}
\alias{has_header}
\title{Ensure the first line of a file has contains all of the expected columns.}
\usage{
has_header(filename, expected_columns)
}
\arguments{
\item{filename}{A file that may or may not have a header}

\item{expected_columns}{A character vector of one or more strings that should appear in a header}
}
\value{
TRUE is the header looks OK, FALSE otherwise
}
\description{
The order is not checked and the file may contain additional columns.
}
