% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylostratr.R
\name{homolog_tree_to_phylostrata}
\alias{homolog_tree_to_phylostrata}
\title{Convert a homolog tree to a phylostrata table}
\usage{
homolog_tree_to_phylostrata(x, FUN, ...)
}
\arguments{
\item{x}{Tree of binary homology results}

\item{FUN}{A function of a a binary homology tree and ... that produces a
phylostrata table}

\item{...}{Additional arguments passed to FUN}
}
\value{
A phylostrata table
}
\description{
Convert a homolog tree to a phylostrata table
}
