% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{is_valid_strata}
\alias{is_valid_strata}
\title{Assert a Strata object is valid, die on failure}
\usage{
is_valid_strata(strata, required = NULL, check_focal = TRUE)
}
\arguments{
\item{strata}{Strata object}

\item{required}{A character vector of required data fields}

\item{check_focal}{logical, if TRUE, then presence of the focal species in
the tree is checked}
}
\description{
Assert a Strata object is valid, die on failure
}
