% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{leafs}
\alias{leafs}
\alias{leafs.Strata}
\alias{leafs.phylo}
\title{Get leaf indices}
\usage{
leafs(x, ...)

\method{leafs}{Strata}(x, ...)

\method{leafs}{phylo}(x, byname = FALSE, ...)
}
\arguments{
\item{x}{phylo object}

\item{...}{Arguments passed on to \code{leafs.phylo}}

\item{byname}{logical Return leaf labels, rather than indices}
}
\description{
Get leaf indices
}
