% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lineage}
\alias{lineage}
\title{Find the ancestors of a node}
\usage{
lineage(tree, id)
}
\description{
Find the ancestors of a node
}
\examples{
set.seed(42)
a <- rtree(10)
a$tip.label <- paste0("t", 1:ntip(a))
a$node.label <- paste0("n", 1:a$Nnode + ntip(a))
}
