% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lineage_to_ancestor_tree}
\alias{lineage_to_ancestor_tree}
\title{Make a tree of ancestors from a lineage}
\usage{
lineage_to_ancestor_tree(lineage)
}
\arguments{
\item{lineage}{A list of ancestors, ranked from old to young}
}
\value{
phylo object with 0 length branches
}
\description{
Make a tree of ancestors from a lineage
}
