% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_building.R
\name{lineages_to_phylo}
\alias{lineages_to_phylo}
\title{Make a tree from the union of lineages}
\usage{
lineages_to_phylo(lineages, remove_subspecies = FALSE)
}
\arguments{
\item{lineages}{list of lineages, where each lineage is a data.frame with
the columns id and name}

\item{remove_subspecies}{Should leafs with subtrees be removed. Currently if
one lineage uses a specific species as a leaf and a second lineage uses a
child of the species, a malformed tree is created where one node is both a
tip and a node.}
}
\value{
phylo object with node names and no branch lengths
}
\description{
Make a tree from the union of lineages
}
