% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{make_blast_database}
\alias{make_blast_database}
\title{Build a blast database for one species}
\usage{
make_blast_database(fastafile, blastdb = "blastdb", verbose = FALSE)
}
\arguments{
\item{fastafile}{The path to a protein FASTA file}

\item{blastdb}{The directory in which the blast database should be stored}

\item{verbose}{Print progress messages}
}
\value{
path to blast database (without extension)
}
\description{
Build a blast database for one species
}
