% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{make_diamond_database}
\alias{make_diamond_database}
\title{Build a diamond database for one species
author urmi}
\usage{
make_diamond_database(fastafile, blastdb = "blastdb", verbose = FALSE)
}
\arguments{
\item{fastafile}{The path to a protein FASTA file}

\item{blastdb}{The directory in which the diamond database should be stored}

\item{verbose}{Print progress messages}
}
\value{
path to diamond database (without extension)
}
\description{
Build a diamond database for one species
author urmi
}
