% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{make_do_if_over}
\alias{make_do_if_over}
\title{Make a filter}
\usage{
make_do_if_over(n = 3, fun = take_first)
}
\arguments{
\item{n}{The number of taxa that must be present before resorting to
\code{fun}}

\item{fun}{The filter function to use for strata with more than \code{n}
representatives}
}
\value{
A filter that can be used in \code{uniprot_cousin_proteomes}
}
\description{
Make a filter
}
