% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{make_jump_matrix}
\alias{make_jump_matrix}
\title{Get comparisons of inferences across scoring systems}
\usage{
make_jump_matrix(d, labels = NULL)
}
\arguments{
\item{d}{data.frame containing a column of group labels and a column of indices}

\item{labels}{character vector of labels}
}
\value{
matrix
}
\description{
Get comparisons of inferences across scoring systems
}
