% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{make_matrix_from_two_strata}
\alias{make_matrix_from_two_strata}
\title{Make a count matrix from two strata}
\usage{
make_matrix_from_two_strata(d)
}
\arguments{
\item{d}{data.frame with two columns holding stratum labels}
}
\value{
CountMatrix object
}
\description{
Make a count matrix from two strata
}
