% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_obo_pdf}
\alias{make_obo_pdf}
\title{Make PDf with multiple obo plots per page}
\usage{
make_obo_pdf(d, file = "obo.pdf", width = 1, height = 5, ...)
}
\arguments{
\item{d}{Besthits table}

\item{file}{Filename for output PDF}

\item{width}{Integer number of plots per row}

\item{height}{Integer number of plots per column}

\item{...}{Additional arguments for \code{plot_obo}}
}
\description{
Make PDf with multiple obo plots per page
}
