% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{make_significance_matrices}
\alias{make_significance_matrices}
\title{Get comparisons of inferences across scoring systems}
\usage{
make_significance_matrices(m)
}
\arguments{
\item{m}{data.frame of mrca}
}
\value{
data.frame of stratum labels
}
\description{
Get comparisons of inferences across scoring systems
}
