% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{make_TIPS_comparison_matrix}
\alias{make_TIPS_comparison_matrix}
\title{Build a matrix relating 2014 study to 2018}
\usage{
make_TIPS_comparison_matrix(strata, results, classifier)
}
\arguments{
\item{strata}{Strata object}

\item{results}{data.frame of highest scoring hits of each gene against each target}

\item{classifier}{function of 'results' that returns a logical vector}

\item{return}{matrix}
}
\description{
With 2014 classifications on the y axis and 2018 on the x axis.
}
