% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{make_tree_relative_to}
\alias{make_tree_relative_to}
\title{Organize the tips with the focal_id on tip}
\usage{
make_tree_relative_to(tree, focal_id)
}
\arguments{
\item{tree}{phylo object}

\item{focal_id}{The name of the taxon to be ordered relative to}
}
\description{
Organize the tips with the focal_id on tip
}
