% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{map_ids}
\alias{map_ids}
\title{Map indices from a to b based off tip labels}
\usage{
map_ids(a, b)
}
\arguments{
\item{a}{phylo object}

\item{b}{phylo object}
}
\value{
matrix mapping indices from `a` to `b`
}
\description{
Map indices from a to b based off tip labels
}
