% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{map_node_label}
\alias{map_node_label}
\title{Map nodes names from a to b}
\usage{
map_node_label(a, b)
}
\arguments{
\item{a}{phylo object with node labels}

\item{b}{phylo object}
}
\description{
The trees a and by must have the same topology and same tip labels.
}
