% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{merge_besthits}
\alias{merge_besthits}
\title{Build a single data.frame with an MRCA column from stratified blast results}
\usage{
merge_besthits(besthits_strata)
}
\arguments{
\item{besthits_strata}{A named list of lists of dataframes. The names are
the phylostrata names. The data.frames are filtered BLAST results.}
}
\value{
A single dataframe holding the top hits of each focal gene against
each subject species.
}
\description{
Build a single data.frame with an MRCA column from stratified blast results
}
