% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{merge_phylo}
\alias{merge_phylo}
\title{Merge fully named subtrees according to a reference tree}
\usage{
merge_phylo(tree, subtrees)
}
\arguments{
\item{tree}{phylo object}

\item{subtrees}{A list of phylo objects. All nodes and leafs of each tree
must be named. All names must be from 'tree'.}
}
\value{
phylo object
}
\description{
Merge fully named subtrees according to a reference tree
}
\examples{
data(atree)
subtrees <- list(
  subtree(atree, 'n19'),
  subtree(atree, 'n18'),
  subtree(atree, 'n15')
)
merge_phylo(atree, subtrees)
}
