% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{ncbi_ancestors}
\alias{ncbi_ancestors}
\title{Get the lineage of a given species}
\usage{
ncbi_ancestors(taxid, byname = FALSE)
}
\arguments{
\item{taxid}{a single NCBI taxon}

\item{byname}{Get ancestor names, instead of NCBI ids}
}
\description{
Get the lineage of a given species
}
