% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-taxonomy.R
\name{ncbi_aunts}
\alias{ncbi_aunts}
\title{Get all ancestral sisters of a taxon}
\usage{
ncbi_aunts(taxid)
}
\arguments{
\item{taxid}{a single NCBI taxon}
}
\value{
phylo object with all NCBI sisters of all ancestors
}
\description{
Get all ancestral sisters of a taxon
}
