% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{ncbi_cousins}
\alias{ncbi_cousins}
\title{Get the species level cousins of a taxid}
\usage{
ncbi_cousins(taxid)
}
\arguments{
\item{taxid}{a single NCBI taxon}
}
\description{
Get the species level cousins of a taxid
}
