% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-taxonomy.R
\name{ncbi_tree}
\alias{ncbi_tree}
\title{Build a phylogenetic tree from a list of NCBI taxon IDs}
\usage{
ncbi_tree(taxa)
}
\arguments{
\item{taxa}{vactor of NCBI taxon IDs}
}
\value{
phylo object
}
\description{
Build a phylogenetic tree from a list of NCBI taxon IDs
}
\examples{
\dontrun{
# raises warning about distances equaling zero, don't worry about it
ncbi_tree(c(9606, 9598, 9593))
}
}
