% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{nleafs}
\alias{nleafs}
\alias{nleafs.Strata}
\alias{nleafs.phylo}
\title{Get the number of leafs}
\usage{
nleafs(x)

\method{nleafs}{Strata}(x)

\method{nleafs}{phylo}(tree)
}
\arguments{
\item{x}{phylo object}
}
\description{
Get the number of leafs
}
