% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{nodes}
\alias{nodes}
\title{Get node indices}
\usage{
nodes(tree, rebase = FALSE, byname = FALSE)
}
\arguments{
\item{tree}{phylo object}

\item{rebase}{logical Make indices base 1}

\item{byname}{logical Return node labels, rather than indices}
}
\description{
Get node indices
}
