% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{normalize_matrix_by_col}
\alias{normalize_matrix_by_col}
\title{Normalize a matrix such that columns sum to 1}
\usage{
normalize_matrix_by_col(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
matrix
}
\description{
Normalize a matrix such that columns sum to 1
}
