% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{organelle_table}
\alias{organelle_table}
\title{Build a table of organelle count statistics}
\usage{
organelle_table(strata, ...)
}
\arguments{
\item{strata}{Strata object}

\item{...}{Additional arguments passed to \code{add_organelle_proteins}}
}
\value{
data.frame with the fields 'species', 'n_mitochondrial', 'n_chloroplast', 'mrca', and 'ps'.
}
\description{
Build a table of organelle count statistics
}
