% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{partial_id_to_name}
\alias{partial_id_to_name}
\title{Convert vector of mixed IDs and names to a vector of names}
\usage{
partial_id_to_name(x)
}
\arguments{
\item{x}{character vector}
}
\value{
Vector of names
}
\description{
Convert vector of mixed IDs and names to a vector of names
}
