% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylostratr.R
\docType{package}
\name{phylostratr}
\alias{phylostratr}
\alias{phylostratr-package}
\title{phylostratr: Phylostratigraphy execution and analysis}
\description{
This is a stub
}
\section{prokaryote_sample}{


A selection of bacterial and archeael species which have UniProt reference
proteoms. These species were selected by randomly choosing one
representative from each class (using the \code{uniprot_sample_prokaryotes}
function with random seed 42).

This sample is used as the root stratum in both the arabidopsis and yeast
vignettes.
}

