% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.Strata}
\alias{plot.Strata}
\title{Plot a Strata object}
\usage{
\method{plot}{Strata}(x, ...)
}
\arguments{
\item{x}{Strata object}

\item{...}{Additional arguments sent to ggtree}
}
\description{
Plot a Strata object
}
