% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_compare_strata}
\alias{plot_compare_strata}
\title{Compare strata between models}
\usage{
plot_compare_strata(x)
}
\arguments{
\item{x}{data.frame with phylostrata classifications as columns}
}
\value{
ggplot heatmap object
}
\description{
Compare strata between models
}
