% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_compare_two_strata}
\alias{plot_compare_two_strata}
\title{Make a heatmap comparing two functions}
\usage{
plot_compare_two_strata(x, normalize_with = function(d) t(apply(d, 1,
  function(x) x/sum(x))), ...)
}
\arguments{
\item{x}{data.frame with two columns of factors corresponding to genewise
assignments across two studies}

\item{normalize_with}{A function for normalizing a count matrix}

\item{...}{Arguments passed to \code{plot_heatmap_with_counts}}
}
\description{
Make a heatmap comparing two functions
}
