% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap_with_counts}
\alias{plot_heatmap_with_counts}
\title{Function for making a heatmap plot with count text}
\usage{
plot_heatmap_with_counts(m, cnt, labels = names(m)[1:2],
  value_trans = function(x) ifelse(x == 0, NA, x),
  scheme = ggplot2::scale_fill_gradient(low = "grey", high = "red"))
}
\arguments{
\item{m}{data.frame where the first two columns contain factors for heatmap
labels and the third column contains proportion.}

\item{cnt}{like m but containing count data}

\item{labels}{for the x and y axis, respectively}

\item{value_trans}{A transformation of the count values, by default, just replace 0 with NA}

\item{scheme}{The aesthetics of the plot, either "myway" or the "highway"}
}
\value{
ggplot object
}
\description{
Function for making a heatmap plot with count text
}
