% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_obo}
\alias{plot_obo}
\title{Plot the max hit scores against each species}
\usage{
plot_obo(d, ...)
}
\arguments{
\item{d}{data.frame of the best hits of focal genes against subject species.
It must have the columns [staxid, qseqid, evalue, score, mrca, ps]}

\item{...}{Additional arguments sent to .arrange_hits}
}
\value{
A named list of ggplot2 objects, with names being the qseqid
}
\description{
Plot the max hit scores against each species
}
