% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_one_obo}
\alias{plot_one_obo}
\title{Plot the max hit scores against each species for one gene}
\usage{
plot_one_obo(stats, qseqid, xlines)
}
\arguments{
\item{stats}{A data.frame with the columns [index, score, eval_bins]}

\item{qseqid}{The query gene ID}

\item{xlines}{A table specifying where the phylostrata lines should be drawn}
}
\value{
ggplot2 object
}
\description{
This is an internal method, a user would normally use \code{plot_obo} or
\code{make_obo_pdf}.
}
