% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_phyloheatmap}
\alias{plot_phyloheatmap}
\title{Build a phylostratigraphic heatmap}
\usage{
plot_phyloheatmap(
  hits,
  tree = NULL,
  n = 50,
  focal_id = NULL,
  to_name = TRUE,
  ...
)
}
\arguments{
\item{hits}{A table of best hits}

\item{tree}{An optional phylo object, if it is not given, a tree will be
infrerred from the taxon IDs in the 'staxid' column. This will work only if
all these IDs are valid NCBI taxonomy IDs.}

\item{n}{The number of genes to display per page}

\item{focal_id}{The focal taxonomy ID (if given, this will be used to order
the tree relative with the focal species on top)}

\item{to_name}{If TRUE, then the tip labels will be converted from taxonomy
IDs to scientific names}
}
\value{
ggtree object
}
\description{
Build a phylostratigraphic heatmap
}
