% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_proteome_stats}
\alias{plot_proteome_stats}
\title{Plot summary statistics for all proteomes}
\usage{
plot_proteome_stats(strata)
}
\arguments{
\item{strata}{Strata object}
}
\description{
Plot summary statistics for all proteomes
}
