% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.Strata}
\alias{print.Strata}
\title{Strata print generic function}
\usage{
\method{print}{Strata}(x, ...)
}
\arguments{
\item{x}{Strata object}

\item{...}{Additional arguments (unused)}
}
\description{
Strata print generic function
}
