% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{proteome_report_table}
\alias{proteome_report_table}
\title{Generate a summary table of proteome stats}
\usage{
proteome_report_table(strata)
}
\arguments{
\item{strata}{Strata object}
}
\description{
Generate a summary table of proteome stats
}
