% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{proteome_stats_table}
\alias{proteome_stats_table}
\title{Build a table of proteome stratistics}
\usage{
proteome_stats_table(strata)
}
\arguments{
\item{strata}{Strata object}
}
\value{
data.frame with the fields 'species', 'protein_length', 'mrca', 'ps', and 'index'.
}
\description{
Build a table of proteome stratistics
}
