% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{prune_phylostrata}
\alias{prune_phylostrata}
\title{Given a strata table, remove given strata}
\usage{
prune_phylostrata(strata_table, strata_names)
}
\arguments{
\item{strata_table}{data.frame with the factored column 'mrca_name'}

\item{strata_names}{character vector of strata to drop}
}
\description{
The genes in a stratum that is removed will be reassigned to the parent
stratum. If the stratum is the root stratum, the genes will instead be
reassigned to the child stratum.
}
\details{
Note that removing strata will result in non-monophyletic branches.
}
