% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{query_sparql}
\alias{query_sparql}
\title{Execute a SPARQL query}
\usage{
query_sparql(template, macros)
}
\arguments{
\item{template}{filename A SPARQL query file that may contain variables that will be replaced with values from `macros`}

\item{macros}{named vector where names are macros that occur in the template SPARQL query and values are what will replace the macro}
}
\value{
table containing the rows of data returned from the query
}
\description{
Execute a SPARQL query
}
