% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{read_blast}
\alias{read_blast}
\title{Read a blast result of the form expected by phylostratr}
\usage{
read_blast(x, with_taxid = TRUE)
}
\arguments{
\item{x}{filename}

\item{with_taxid}{If TRUE, expect the staxid column to be in the table}
}
\description{
Read a blast result of the form expected by phylostratr
}
