% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{rename_species}
\alias{rename_species}
\title{Rename a species in a Strata object}
\usage{
rename_species(strata, old_name, new_name)
}
\arguments{
\item{strata}{Strata object}

\item{old_name}{character vector of names that are currently in the Strata object}

\item{new_name}{character vector of names to replace the old_name vector}
}
\value{
Strata object with the same number of species
}
\description{
Rename a species in a Strata object
}
\examples{
\dontrun{
data(saccharomyces)
x <- rename_species(saccharomyces, "Saccharomyces_arboricola", "Sa")
is_valid_strata(x)

x <- rename_species(saccharomyces,
  c("Saccharomyces_arboricola", "Saccharomyces_cerevisiae"), c("Sa", "Sc"))
is_valid_strata(x)
}
}
