% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{replace_branch}
\alias{replace_branch}
\title{Replace a branch on a Strata}
\usage{
replace_branch(x, y, node)
}
\arguments{
\item{x}{Strata object}

\item{y}{Strata object}

\item{node}{The name of a node in x where y will be placed}
}
\description{
Replace a branch on a Strata
}
\examples{
data(saccharomyces)
x = saccharomyces
y = subtree(saccharomyces, 's2')
y = strata_convert(y, target='tip', to='name')
x = replace_branch(x, y, "s2")
}
