% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylostratr.R
\name{result_tree_to_homolog_tree}
\alias{result_tree_to_homolog_tree}
\title{Convert tree of p-values to a homolog tree}
\usage{
result_tree_to_homolog_tree(x, FUN, ...)
}
\arguments{
\item{x}{Tree of raw results for pairwise proteome comparison}

\item{FUN}{A function of one branch of a tree that assigns a binary value to each leaf}

\item{...}{Additional arguments passed to FUN}
}
\value{
A tree with binary leaves
}
\description{
Convert tree of p-values to a homolog tree
}
