% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylostratr.R
\name{result_tree_to_pvalue_tree}
\alias{result_tree_to_pvalue_tree}
\title{Convert tree of raw results to a p-value tree}
\usage{
result_tree_to_pvalue_tree(x, FUN, ...)
}
\arguments{
\item{x}{Tree of raw results for pairwise comparisons}

\item{FUN}{A function of one branch of a tree that assigns a p-value to each leaf}

\item{...}{Additional arguments passed to FUN}
}
\value{
A p-value tree
}
\description{
Convert tree of raw results to a p-value tree
}
