% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{run_hmmscan}
\alias{run_hmmscan}
\title{Find domains in a protein file}
\usage{
run_hmmscan(fastafile, db, nthreads = 1)
}
\arguments{
\item{fastafile}{The name of a protein FASTA file}

\item{db}{The name of a HMM database (e.g. PFAM)}

\item{nthreads}{Number of cores to use (leave two or more threads free)}
}
\value{
A list containing two dataframe: tblout and domtblout (see HMMer manual)
}
\description{
Find domains in a protein file
}
