% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saccharomyces}
\alias{saccharomyces}
\title{A custom Strata object for the Saccharomyces genus}
\format{Strata}
\usage{
saccharomyces
}
\description{
A custom Strata object for the Saccharomyces genus
}
\keyword{datasets}
