% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{sisters}
\alias{sisters}
\alias{sisters.Strata}
\alias{sisters.phylo}
\title{Get the sisters of a node}
\usage{
sisters(x, id, ...)

\method{sisters}{Strata}(x, id, ...)

\method{sisters}{phylo}(x, id, type = "auto", ...)
}
\arguments{
\item{x}{phylo object}

\item{id}{vector of ids or names}

\item{...}{Arguments sent to \code{sisters.phylo}}

\item{type}{id type: ['name', 'index', 'auto']}
}
\value{
integer vector of indices
}
\description{
Get the sisters of a node
}
