% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{sort_strata}
\alias{sort_strata}
\title{Sort strata relative to focal species}
\usage{
sort_strata(strata)
}
\arguments{
\item{strata}{Strata object}
}
\value{
Strata object with reorded tips
}
\description{
Sort strata relative to focal species
}
