% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylostratr.R
\name{species_to_species_tree}
\alias{species_to_species_tree}
\title{Get a species tree tracing back to origin for the focal species}
\usage{
species_to_species_tree(x, FUN, ...)
}
\arguments{
\item{x}{The focal species taxon ID}

\item{FUN}{A function of x and ... that builds a species tree}

\item{...}{Additional arguments passed to FUN}
}
\value{
A tree with species leaves and optional node names
}
\description{
Get a species tree tracing back to origin for the focal species
}
