% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{standardize_strata}
\alias{standardize_strata}
\title{Standardize two or more phylostrata tables}
\usage{
standardize_strata(strata_tables)
}
\arguments{
\item{strata_tables}{phylostrata tables, all must the column 'mrca_name'}
}
\value{
list of standardized phylostrata tables
}
\description{
Each table is subset with the union of sequence IDs. Also all strata that
are not shared between all input tables are dropped, with the genes
reassigned to older strata as needed.
}
