% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{strata2mrca}
\alias{strata2mrca}
\title{Get a map from taxid id to MRCA and stratum level}
\usage{
strata2mrca(strata)
}
\arguments{
\item{strata}{A three-level list. The first level has one element for each
node in the focal species lineage (and is named accordingly). The second
level has one element for each 'uncle'. The third level is a possibly empty
vector of taxon IDs.}
}
\description{
Get a map from taxid id to MRCA and stratum level
}
