% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{strata_apply}
\alias{strata_apply}
\title{Apply f to each outgroup branch ascending node id}
\usage{
strata_apply(strata, f, id = strata@focal_id, ...)
}
\arguments{
\item{strata}{Strata object}

\item{f}{A function of a phylo object that return a phylo object}

\item{id}{A name or index}

\item{...}{Additional arguments passed to f}
}
\value{
Strata object
}
\description{
Apply f to each outgroup branch ascending node id
}
