% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{strata_besthits}
\alias{strata_besthits}
\title{Load each blast result and filter out the best hit against each query gene}
\usage{
strata_besthits(strata)
}
\arguments{
\item{strata}{Strata object with 'blast_result' vector in data
of a possibly empty list of filenames. The filenames are raw BLAST results.}
}
\value{
Strata object with 'besthits' field in data slot. This field holds a
data.frame for each target species, where each data.frame is a filtered
blast result
}
\description{
Load each blast result and filter out the best hit against each query gene
}
