% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{strata_blast}
\alias{strata_blast}
\title{Blast uncle-level strata}
\usage{
strata_blast(query, strata, makedb_args = list(), blast_args = list())
}
\arguments{
\item{query}{The protein FASTA file for the focal species}

\item{strata}{Uncle level strata}

\item{makedb_args}{Additional arguments passed to \code{make_blast_database}}

\item{blast_args}{Additional arguments passed to \code{run_blastp}}
}
\description{
Takes a uncle-level strata, blasts each leaf, returns a flat strata with blast result filenames
}
