% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{strata_convert}
\alias{strata_convert}
\title{Convert strata data, tip, and node names to and from NCBI taxonomy IDs}
\usage{
strata_convert(strata, target = "tip", to = "id")
}
\arguments{
\item{strata}{Strata object}

\item{target}{What to convert, may be 'tip', 'node', or 'all'}

\item{to}{What to convert to, may be 'id' or 'name'}
}
\value{
Strata object with new names
}
\description{
Convert strata data, tip, and node names to and from NCBI taxonomy IDs
}
