% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{strata_fold}
\alias{strata_fold}
\title{Apply f to each outgroup branch ascending node id, returning a list}
\usage{
strata_fold(strata, f, ...)
}
\arguments{
\item{strata}{Strata object}

\item{f}{A function of a phylo object that may return anything}

\item{...}{Additional arguments passed to f}
}
\value{
A named list, with names corresponding to phylostrata
}
\description{
Apply f to each outgroup branch ascending node id, returning a list
}
