% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{strata_from_blast_dir}
\alias{strata_from_blast_dir}
\title{Build a Strata object from a directory of BLAST results}
\usage{
strata_from_blast_dir(focal_species, blastdir = ".", ext = "tab", ...)
}
\arguments{
\item{focal_species}{character: the taxonomy ID of the species that was used in the BLAST}

\item{blastdir}{filename: the directory containing the BLAST results}

\item{ext}{character: the file extension (e.g. 'tab' or 'blast.tab')}

\item{...}{additional arguments sent to \code{strata_from_taxids}}
}
\value{
Strata object wrapping a NCBI tree built from the target species.
}
\description{
The directory must contain one file for each search of the focal species
against a target species. The format of the filenames must be
<taxon_id>.<ext>. The species must all have NCBI taxonomy IDs. The files
must have headers and the columns expected by \code{read_blast}.
}
