% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{strata_from_taxids}
\alias{strata_from_taxids}
\title{Make a Strata object from a list of taxon ids}
\usage{
strata_from_taxids(focal_species, taxids, clean = TRUE)
}
\arguments{
\item{focal_species}{character: the taxonomy ID of the species that was used in the BLAST}

\item{taxids}{character: list of NCBI taxonomy IDs}

\item{clean}{logical: Should leafs with descendents be removed? This can occur when
both a species and a descendent subspecies are in the lineage set.}
}
\value{
Strata object
}
\description{
Make a Strata object from a list of taxon ids
}
