% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{strata_uniprot_pfam_map}
\alias{strata_uniprot_pfam_map}
\title{Retrieve uniprot to pfam maps for all taxa}
\usage{
strata_uniprot_pfam_map(strata, dir = "pfam")
}
\arguments{
\item{strata}{Strata object}

\item{dir}{director to which maps should be written}
}
\value{
Strata object with uniprot2pfam field in data, which stores the
        filenames of the retrieved maps.
}
\description{
Retrieve uniprot to pfam maps for all taxa
}
