% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{subset_fasta}
\alias{subset_fasta}
\title{Select ids from a FASTA file, creating a new file}
\usage{
subset_fasta(f, ids, prefix = "zzz_")
}
\arguments{
\item{f}{filename of FASTA protein sequence file}

\item{ids}{a vector for subseting (either logical with one element for each
gene, or integers with 1 <= i <= length(seqs).}

\item{prefix}{a file basename prefix for the output FASTA file}
}
\value{
filename
}
\description{
Select ids from a FASTA file, creating a new file
}
