% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{subset_phylo}
\alias{subset_phylo}
\title{Select a subset of nodes from a tree}
\usage{
subset_phylo(tree, id, collapse = TRUE, descend = TRUE)
}
\arguments{
\item{tree}{phylo object}

\item{id}{vector of ids or names}

\item{collapse}{logical collapse edges around nodes with a single descendent}

\item{descend}{logical include all descendents of each id}
}
\value{
phylo object
}
\description{
Select a subset of nodes from a tree
}
\examples{
data(atree)
subset_phylo(atree, c('n15', 'n19'))
subset_phylo(atree, c('t7', 't4', 't1'))
}
