% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{t.CountMatrix}
\alias{t.CountMatrix}
\title{Transpose a CountMatrix}
\usage{
\method{t}{CountMatrix}(x)
}
\arguments{
\item{x}{CountMatrix object}
}
\description{
Take the traspose of the matrix and swap the labels
}
