% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{take_first}
\alias{take_first}
\title{Select representatives for a strata}
\usage{
take_first(x)
}
\arguments{
\item{x}{A list of uncles at a specific stratum, each of which contains a
list of descendent taxids}
}
\description{
Select representatives for a strata
}
