% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxid2name}
\alias{taxid2name}
\title{Get the scientific names for a list of taxids}
\usage{
taxid2name(taxids, ...)
}
\arguments{
\item{taxids}{A list of NCBI taxonomy ids}

\item{...}{Arguments passed to \code{taxid2summary}}
}
\description{
Get the scientific names for a list of taxids
}
