% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxid2summary}
\alias{taxid2summary}
\title{Transform list of taxids to list of taxid summaries}
\usage{
taxid2summary(taxids, chunk_size = 250)
}
\arguments{
\item{taxids}{Vector of NCBI taxonomy ids}

\item{chunk_size}{number of records to retrive at a time (the default should be fine)}
}
\description{
Transform list of taxids to list of taxid summaries
}
