% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{thin_fasta}
\alias{thin_fasta}
\title{Select n entries from a FASTA file, creating a new file}
\usage{
thin_fasta(f, n, ...)
}
\arguments{
\item{f}{filename of FASTA protein sequence file}

\item{n}{integer number of entries to extract}

\item{...}{arguments passed to subset_fasta}
}
\value{
filename
}
\description{
Selects a total of n entries from a FASTA file by selecting every entry where
}
\details{
i % n == n / 2

Where i is the element index (1,2, ... N), N is the total number of
sequences in the genome, '%' is the modulo operator, and '/' is integer
division.

The final term 'n / 2' ensures the first and last selected entries are
equidistant (for even values of n) from the beginning and end of the
sequences.
}
