% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tree_size}
\alias{tree_size}
\title{Get the number of species and ancestors in the tree}
\usage{
tree_size(tree)
}
\arguments{
\item{tree}{phylo object}
}
\value{
integer
}
\description{
Get the number of species and ancestors in the tree
}
