% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_add_idmap}
\alias{uniprot_add_idmap}
\title{Extract an ID map from uniprot fasta files}
\usage{
uniprot_add_idmap(strata)
}
\arguments{
\item{strata}{Strata object}
}
\value{
Strata object with 'idmap' data field
}
\description{
This assumes the fasta files respect the UniProt fasta header format.
}
