% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_cousin_genomes}
\alias{uniprot_cousin_genomes}
\title{Retrive the sequences from the uniprot cousins}
\usage{
uniprot_cousin_genomes(cousins, dir = "strata", prefix = "ps_",
  filter = identity, dryrun = FALSE, ...)
}
\arguments{
\item{cousins}{The output of \code{uniprot_cousins}}

\item{dir}{Directory in which to write sequences}

\item{prefix}{The phylostratum-level prefix}

\item{filter}{A function to filter desired target species from the cousin list}

\item{dryrun}{If TRUE, do not download files or create directories}

\item{...}{Additional arguments passed to \code{uniprot_retrieve_genomes}}
}
\value{
Nothing, this function is run for its effects
}
\description{
Retrive the sequences from the uniprot cousins
}
\examples{
\dontrun{
# uniprot_cousins(3702) \%>\% uniprot_cousin_genomes(filter=takefirst)
}
}
