% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_cousins}
\alias{uniprot_cousins}
\title{Given a focal taxid, build a tree of}
\usage{
uniprot_cousins(taxid)
}
\arguments{
\item{taxid}{The focal species NCBI taxon id}
}
\value{
Strata object
}
\description{
Given a focal taxid, build a tree of
}
