% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_map2pfam}
\alias{uniprot_map2pfam}
\title{Map UniProt IDs for an organism to PFAM IDs}
\usage{
uniprot_map2pfam(taxid)
}
\arguments{
\item{taxid}{Species NCBI taxonomy ID}
}
\value{
A data.frame with columns 'uniprotID' and 'pfamID'
}
\description{
Map UniProt IDs for an organism to PFAM IDs
}
