% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_organelle_ids}
\alias{uniprot_organelle_ids}
\title{Get the uniprot ids for organelle proteins of a given taxid}
\usage{
uniprot_organelle_ids(taxid, organelle = "Mitochondrion")
}
\arguments{
\item{taxid}{Taxon ID for species of interest}

\item{organelle}{string, one of 'mitochondrion', 'chloroplast'}
}
\value{
A numeric vector of Uniprot ids
}
\description{
Get the uniprot ids for organelle proteins of a given taxid
}
