% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_retrieve_genomes}
\alias{uniprot_retrieve_genomes}
\title{Download UniProt proteomes for given taxa}
\usage{
uniprot_retrieve_genomes(taxa, keep_isoforms = FALSE, dir = "uniprot-seqs",
  dryrun = FALSE)
}
\arguments{
\item{taxa}{Vector of NCBI taxonomy ids}

\item{keep_isoforms}{Should all isoforms of each protein be included?}

\item{dir}{Directory in which to write all FASTA files}

\item{dryrun}{If TRUE, do not download genomes, just print the URLs and destination files}
}
\description{
Each proteom is written to a file with the name '<taxid>.faa', for example,
Arabidopsis thaliana, which has the id 3702, will be written the '3702.faa'.
}
\examples{
\dontrun{
# uniprot_downstream_ids(3701) \%>\% uniprot_retrieve_genomes
}
}
