% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_retrieve_proteome_table}
\alias{uniprot_retrieve_proteome_table}
\title{Retrieve UniProt proteome table}
\usage{
uniprot_retrieve_proteome_table(taxid)
}
\arguments{
\item{taxid}{An NCBI taxonomy id}
}
\description{
Retrieve all proteins in a "representative" proteome for the given species.
}
\examples{
\dontrun{
x <- uniprot_retrieve_proteome_table(3702)
}
}
