% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_strata}
\alias{uniprot_strata}
\title{Given a focal taxid, build a UniProt-based Strata object}
\usage{
uniprot_strata(taxid, from = 2, drop.names = NULL)
}
\arguments{
\item{taxid}{The focal species NCBI taxon id}

\item{from}{stratum to begin from, where 1 is 'cellular organisms'}
}
\value{
Strata object
}
\description{
Given a focal taxid, build a UniProt-based Strata object
}
