% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_uncle_genomes}
\alias{uniprot_uncle_genomes}
\title{Retrive the sequences from the uniprot uncle ids}
\usage{
uniprot_uncle_genomes(uncle_list, dir = "strata", prefix = "ps_", ...)
}
\arguments{
\item{dir}{Directory in which to write sequences}

\item{prefix}{The phylostratum-level prefix}

\item{uncle_id}{The output of \code{uniprot_uncle_ids}}
}
\description{
Retrive the sequences from the uniprot uncle ids
}
