% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_weight_by_ref}
\alias{uniprot_weight_by_ref}
\title{Make reference-species preferring weight vector for diverse_subtree}
\usage{
uniprot_weight_by_ref(weight = 1.05, clade = 2759)
}
\arguments{
\item{weight}{How much to prefer the reference species. The default, 1.05,
will weakly prefer them, acting mostly as a tie-breaker. Higher weights
could lead to reduced diversity.}

\item{clade}{Weight taxa descending from this clade. The default is 2759
(Eukaryota).}
}
\description{
Make reference-species preferring weight vector for diverse_subtree
}
