% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prokaryotes.R
\name{use_custom_prokaryote_tree}
\alias{use_custom_prokaryote_tree}
\title{Use this function to add the custom tree to the analysis}
\usage{
use_custom_prokaryote_tree(x, prokaryote.phylo)
}
\arguments{
\item{x}{an existing tree you want to add to, it is expected to contain a eukaryotic branch}

\item{prokaryote.phylo}{a phylo object of the custom prokaryotic tree to add}
}
\value{
tree integrating both the eukaryotic tree from x and the custom prokaryotic tree
}
\description{
Based on use_recommended_prokaryotes but enables using custom trees.
Integrates a custom prokaryotic phylogenetic tree with a eukaryotic tree for
phylostratigraphic analysis.
}
