% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{use_recommended_prokaryotes}
\alias{use_recommended_prokaryotes}
\title{Use the prokaryotic species that I use}
\usage{
use_recommended_prokaryotes(x)
}
\arguments{
\item{x}{Strata object}
}
\value{
x Strata object with the basal stratum (ID=131567) replaced
}
\description{
This set of species is currently not all that great. It is a sample of
bacterial and archaeal species from each class in each domain.
}
