% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vet_phyid}
\alias{vet_phyid}
\title{Convert node/tip name to id, check size}
\usage{
vet_phyid(tree, id, len = NULL)
}
\arguments{
\item{tree}{phylo object}

\item{id}{integer id or a node label}

\item{len}{require id to be of this length (NULL for no requirement)}
}
\value{
on success, return a valid integrel id
}
\description{
Convert node/tip name to id, check size
}
