% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{wrap_uniprot_id_retrieval}
\alias{wrap_uniprot_id_retrieval}
\title{Internal funcion for wrapping ID retrieval from UniProt}
\usage{
wrap_uniprot_id_retrieval(db, query, delay = FALSE, cast = identity)
}
\arguments{
\item{db}{UniProt database to search}

\item{query}{UniProt query string}

\item{delay}{Sleep for 0.3 seconds before retrieving (polite in loops)}

\item{cast}{A function for type casting the resulting IDs (e.g. as_integer)}
}
\value{
vector of IDs or other literals (nothing structured)
}
\description{
Internal funcion for wrapping ID retrieval from UniProt
}
