% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{add_taxa}
\alias{add_taxa}
\title{Add an id to a representative list}
\usage{
add_taxa(strata, taxa)
}
\arguments{
\item{strata}{A Strata object (all IDs MUST be NCBI taxonomy IDs)}

\item{taxa}{NCBI taxon ids to add}
}
\value{
Strata object
}
\description{
Add an id to a representative list
}
