% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{add_taxa}
\alias{add_taxa}
\alias{add_taxa.Strata}
\alias{add_taxa.phylo}
\title{Add an id to a representative list}
\usage{
add_taxa(x, taxa, ...)

\method{add_taxa}{Strata}(x, taxa, ...)

\method{add_taxa}{phylo}(x, taxa, ...)
}
\arguments{
\item{x}{A Strata or phylo object (all IDs MUST be NCBI taxonomy IDs)}

\item{taxa}{NCBI taxon ids to add}

\item{...}{Additional arguments (not used currently)}
}
\value{
Strata or phylo object
}
\description{
Add an id to a representative list
}
