% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{as_named_strata}
\alias{as_named_strata}
\title{Substitute names for taxids in strata}
\usage{
as_named_strata(x, scinames = NULL, depth = 1)
}
\arguments{
\item{x}{named list, where all names are NCBI taxon IDs}

\item{scinames}{named character vector, where names are NCBI taxon IDs and
elements are scientific names}

\item{depth}{nest depth of lists with NCBI taxon ID names}
}
\value{
named list, where all names are scientific names
}
\description{
Substitute names for taxids in strata
}
