% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{as_named_strata}
\alias{as_named_strata}
\title{Substitute names for taxids in strata}
\usage{
as_named_strata(x)
}
\arguments{
\item{x}{strata list, as made by \code{uniprot_cousins}}
}
\description{
Substitute names for taxids in strata
}
