% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{clean_phyid}
\alias{clean_phyid}
\alias{clean_phyid.Strata}
\alias{clean_phyid.phylo}
\title{Convert node/tip name to id, check size}
\usage{
clean_phyid(x, ...)

\method{clean_phyid}{Strata}(x, ...)

\method{clean_phyid}{phylo}(x, id, len = NULL, type = "auto")
}
\arguments{
\item{x}{phylo or Strata object}

\item{id}{integer id or a node label}

\item{len}{require id to be of this length (NULL for no requirement)}

\item{type}{['auto', 'index', or 'name'], is 'index' then the id is expected
to be an index in the phylo object, if 'name' it is expected to be a node or
tip name, if 'auto' the value is assumed to be a name if it is a character
string.}
}
\value{
on success, return a valid integrel id
}
\description{
Convert node/tip name to id, check size
}
