% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{descendent_nodes}
\alias{descendent_nodes}
\title{Get all nodes (leafs and ancestors) descending from, and including, a set of ids}
\usage{
descendent_nodes(tree, id)
}
\arguments{
\item{tree}{phylo object}

\item{id}{vector of ids or names}
}
\value{
integer vector of leaf and node ids
}
\description{
Get all nodes (leafs and ancestors) descending from, and including, a set of ids
}
