% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{descendent_nodes}
\alias{descendent_nodes}
\alias{descendent_nodes.Strata}
\alias{descendent_nodes.phylo}
\title{Get all nodes (leafs and ancestors) descending from, and including, a set of ids}
\usage{
descendent_nodes(x, ...)

\method{descendent_nodes}{Strata}(x, ...)

\method{descendent_nodes}{phylo}(x, id, type = "auto")
}
\arguments{
\item{x}{phylo object}

\item{id}{vector of ids or names}

\item{type}{id type: ['name', 'id', 'auto']}
}
\value{
integer vector of leaf and node ids
}
\description{
Get all nodes (leafs and ancestors) descending from, and including, a set of ids
}
