% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{descendents}
\alias{descendents}
\title{Get all leafs descending from a set of ids}
\usage{
descendents(tree, id, ...)
}
\arguments{
\item{tree}{phylo object}

\item{id}{vector of ids or names}

\item{...}{Arguments sent to descendent_nodes}
}
\value{
integer vector of leaf ids
}
\description{
This is a simple wrapper around \code{descendent_nodes}.
}
