% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{find_revenants}
\alias{find_revenants}
\title{Find genes that with matches that skip strata}
\usage{
find_revenants(d, classifier = classify_by_evalue(1e-05))
}
\arguments{
\item{d}{data.frame of maximal hits}

\item{classifier}{function of a blast result that infers whether each
subject species contains a homolog}
}
\value{
data.frame with columns [qseqid, ps, mrca, basal_ps, n].  Where
qseqid is the focal species taxond IDs, ps is the phylostratum level (where
1 is generally cellular_organisms) of strata that are skipped, mrca is the
taxon ID of the ancestor, basal_ps is the oldest stratum with a match to
qseqid, and n is the total number of revenants for seqid.
}
\description{
Find genes that with matches that skip strata
}
