% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{get_root}
\alias{get_root}
\alias{get_root.Strata}
\alias{get_root.phylo}
\title{Get the root of a tree}
\usage{
get_root(x, ...)

\method{get_root}{Strata}(x, ...)

\method{get_root}{phylo}(x, ...)
}
\arguments{
\item{x}{phylo object}

\item{...}{Arguments passed to \code{is_root}}
}
\value{
index of the root node (or nodes if there are more than one)
}
\description{
There should normally be only one, but if there are more, this function will
find all of them.
}
