% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{leafs}
\alias{leafs}
\title{Get leaf indices}
\usage{
leafs(tree, byname = FALSE)
}
\arguments{
\item{tree}{phylo object}

\item{byname}{logical Return leaf labels, rather than indices}
}
\description{
Get leaf indices
}
