% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{lineage}
\alias{lineage}
\title{Find the ancestors of a node}
\usage{
lineage(tree, id, use_name = FALSE)
}
\arguments{
\item{tree}{phylo object}

\item{id}{A single name or index}

\item{use_name}{Should names be used rather than internal ids}
}
\description{
Find the ancestors of a node
}
\examples{
data(atree)
lineage(atree, 't1')
lineage(atree, 1)
}
