% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{lineage}
\alias{lineage}
\alias{lineage.Strata}
\alias{lineage.phylo}
\alias{tree_size.phylo}
\title{Find the ancestors of a node}
\usage{
lineage(x, ...)

\method{lineage}{Strata}(x, ...)

\method{lineage}{phylo}(x, id, use_name = FALSE, type = "auto")

\method{tree_size}{phylo}(x)
}
\arguments{
\item{x}{phylo object}

\item{id}{A single name or index}

\item{use_name}{Should names be used rather than internal ids}

\item{type}{id type: ['name', 'id', 'auto']}
}
\description{
Find the ancestors of a node
}
\examples{
data(atree)
lineage(atree, 't1')
lineage(atree, 1)
}
