% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lineages_to_phylo}
\alias{lineages_to_phylo}
\title{Make a tree from the union of lineages}
\usage{
lineages_to_phylo(lineages)
}
\arguments{
\item{lineages}{list of lineages, where each lineage is a data.frame with
the columns id and name}
}
\value{
phylo object with node names and no branch lengths
}
\description{
Make a tree from the union of lineages
}
