% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_building.R
\name{lineages_to_phylo}
\alias{lineages_to_phylo}
\title{Make a tree from the union of lineages}
\usage{
lineages_to_phylo(lineages, clean = FALSE)
}
\arguments{
\item{lineages}{list of lineages, where each lineage is a data.frame with
the columns id and name}

\item{clean}{Should leafs with descendents be removed? This can occur when
both a species and a descendent subspecies are in the lineage set.}
}
\value{
phylo object with node names and no branch lengths
}
\description{
Make a tree from the union of lineages
}
