% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{make_blast_database}
\alias{make_blast_database}
\title{Build a blast database for one species}
\usage{
make_blast_database(taxid, indir = "uniprot-seqs", outdir = "blastdb")
}
\arguments{
\item{taxid}{A single NCBI taxon id}

\item{indir}{The directory in which to find the protein FASTA sequence
(should contain a file '<taxid>.faa')}

\item{outdir}{The directory in which the blast database should be stored}
}
\value{
path to blast database (without extension)
}
\description{
Build a blast database for one species
}
