% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{make_significance_matrices}
\alias{make_significance_matrices}
\title{Get comparisons of inferences across scoring systems}
\usage{
make_significance_matrices(results, cache = "significance_list.Rda")
}
\arguments{
\item{results}{data.frame of highest scoring hits of each gene against each target}

\item{cache}{A file where data are cached}

\item{return}{list of matrices}
}
\description{
Get comparisons of inferences across scoring systems
}
