% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{make_significance_matrices}
\alias{make_significance_matrices}
\title{Get comparisons of inferences across scoring systems}
\usage{
make_significance_matrices(m, cache = "significance_list.Rda")
}
\arguments{
\item{m}{data.frame of mrca}

\item{cache}{A file where data are cached}
}
\value{
list of matrices
}
\description{
Get comparisons of inferences across scoring systems
}
