% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{merge_besthits}
\alias{merge_besthits}
\title{Build a single data.frame with an MRCA column from stratified blast results}
\usage{
merge_besthits(besthits_strata)
}
\arguments{
\item{besthits_strata}{The output of \code{strata_besthits}}
}
\description{
Build a single data.frame with an MRCA column from stratified blast results
}
