% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{nodes}
\alias{nodes}
\alias{nodes.Strata}
\alias{nodes.phylo}
\title{Get node indices}
\usage{
nodes(x, ...)

\method{nodes}{Strata}(x, ...)

\method{nodes}{phylo}(x, rebase = FALSE, byname = FALSE)
}
\arguments{
\item{x}{phylo object}

\item{rebase}{logical Make indices base 1}

\item{byname}{logical Return node labels, rather than indices}
}
\description{
Get node indices
}
