% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{parent}
\alias{parent}
\alias{parent.Strata}
\alias{parent.phylo}
\title{Vectorized node parent getter}
\usage{
parent(x, ...)

\method{parent}{Strata}(x, ...)

\method{parent}{phylo}(x, id = 1:tree_size(x), type = "auto")
}
\arguments{
\item{x}{phylo object}

\item{id}{vector of ids or names}

\item{type}{id type: ['name', 'id', 'auto']}
}
\value{
vector of parent ids (NA if root)
}
\description{
Vectorized node parent getter
}
