% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_obo}
\alias{plot_obo}
\title{Plot the max hit scores against each species}
\usage{
plot_obo(d, lower.bound = 0, upper.bound = 100)
}
\arguments{
\item{d}{data.frame of the best hits of focal genes against subject species.
It must have the columns [staxid, qseqid, evalue, score, mrca, ps]}

\item{lower.bound}{A lower bound at which to truncate scores}

\item{upper.bound}{An upper bound at which to truncate scores}
}
\value{
A named list of ggplot2 objects, with names being the qseqid
}
\description{
Plot the max hit scores against each species
}
