% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_proteome_lengths}
\alias{plot_proteome_lengths}
\title{Plot the ordered lengths of all proteins in all proteomes}
\usage{
plot_proteome_lengths(strata)
}
\arguments{
\item{strata}{Strata object}
}
\value{
ggplot object
}
\description{
Plot the ordered lengths of all proteins in all proteomes
}
