% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{prune}
\alias{prune}
\alias{prune.Strata}
\alias{prune.phylo}
\title{Remove the specified indices, and their descendents, from the tree}
\usage{
prune(x, id, ...)

\method{prune}{Strata}(x, id, ...)

\method{prune}{phylo}(x, id, type = "auto", ...)
}
\arguments{
\item{x}{phylo object}

\item{id}{vector of indices or names}

\item{...}{Arguments sent to \code{prune.phylo}}

\item{type}{id type: ['name', 'index', 'auto']}
}
\description{
Remove the specified indices, and their descendents, from the tree
}
