% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{prune}
\alias{prune}
\title{Remove the specified indices, and their descendents, from the tree}
\usage{
prune(tree, id)
}
\arguments{
\item{tree}{phylo object}

\item{id}{vector of indices or names}
}
\description{
Remove the specified indices, and their descendents, from the tree
}
