% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{prune}
\alias{prune}
\alias{prune.Strata}
\alias{prune.phylo}
\title{Remove the specified indices, and their descendents, from the tree}
\usage{
prune(x, ...)

\method{prune}{Strata}(x, ...)

\method{prune}{phylo}(tree, id, type = "auto")
}
\arguments{
\item{x}{phylo object}

\item{id}{vector of indices or names}

\item{type}{id type: ['name', 'id', 'auto']}
}
\description{
Remove the specified indices, and their descendents, from the tree
}
