% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{run_blastp}
\alias{run_blastp}
\title{BLAST query protein FASTA file against a target species}
\usage{
run_blastp(query_fastafile, target_taxid, blastdb, blastresult, nthreads = 1,
  seg = FALSE)
}
\arguments{
\item{query_fastafile}{A protein FASTA file for the focal species}

\item{target_taxid}{The target NCBI taxon ID}

\item{blastdb}{The directory in which the blast database should be made}

\item{blastresult}{The output TAB-delimited result file}

\item{nthreads}{Number of threads}

\item{seg}{Whether to mask the query protein sequences}
}
\description{
BLAST query protein FASTA file against a target species
}
