% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{set_node_names}
\alias{set_node_names}
\title{If nodes are not named, give them default names}
\usage{
set_node_names(tree, default_names = paste0("n", nodes(tree)))
}
\arguments{
\item{tree}{phylo object}

\item{default_names}{node names to use if they are missing. Must be a vector
with a number of elements equal to the number of nodes in the tree.}
}
\value{
phylo object
}
\description{
In phylostratr, all nodes in a tree need to be named, since the names are
used to merge trees.
}
