% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{sister_trees}
\alias{sister_trees}
\alias{sister_trees.Strata}
\alias{sister_trees.phylo}
\title{Get list of sister trees for a given node}
\usage{
sister_trees(x, id, ...)

\method{sister_trees}{Strata}(x, id, ...)

\method{sister_trees}{phylo}(x, id, type = "auto")
}
\arguments{
\item{x}{phylo object}

\item{id}{vector of ids or names}

\item{...}{Arguments passed to \code{sister_trees.phylo}}

\item{type}{id type: ['name', 'id', 'auto']}
}
\value{
list of phylo objects
}
\description{
Get list of sister trees for a given node
}
