% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{sisters}
\alias{sisters}
\title{Get the sisters of a node}
\usage{
sisters(tree, id, type = "auto")
}
\arguments{
\item{tree}{phylo object}

\item{id}{vector of ids or names}

\item{type}{id type: ['name', 'id', 'auto']}
}
\value{
integer vector of indices
}
\description{
Get the sisters of a node
}
