% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{strata_besthits}
\alias{strata_besthits}
\title{Load each blast result and filter out the best hit against each query gene}
\usage{
strata_besthits(blast_strata)
}
\arguments{
\item{blast_strata}{A named list of phylostrata where each element consists
of a possibly empty list of filenames. The filenames are raw BLAST results.}
}
\value{
A list of lists of data.frames, where each data.frame is a filtered blast result
}
\description{
Load each blast result and filter out the best hit against each query gene
}
