% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{strata_besthits}
\alias{strata_besthits}
\title{Load each blast result and filter out the best hit against each query gene}
\usage{
strata_besthits(blast_strata)
}
\arguments{
\item{blast_strata}{The output of strata_blast}
}
\value{
A list of lists of filtered blast results
}
\description{
Load each blast result and filter out the best hit against each query gene
}
