% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{strata_blast}
\alias{strata_blast}
\title{Blast strata}
\usage{
strata_blast(strata, makedb_args = list(), blast_args = list())
}
\arguments{
\item{strata}{Strata object where the 'faa' vector is included in the data slot}

\item{makedb_args}{Additional arguments passed to \code{make_blast_database}}

\item{blast_args}{Additional arguments passed to \code{run_blastp}}
}
\value{
named list of phylostrata, where each element is a vector of blast result filenames
}
\description{
Blast strata
}
