% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{subtree}
\alias{subtree}
\alias{subtree.Strata}
\alias{subtree.phylo}
\title{Select a subset of nodes from a tree}
\usage{
subtree(x, id, ...)

\method{subtree}{Strata}(x, id, type = "name", ...)

\method{subtree}{phylo}(x, id, collapse = TRUE, descend = TRUE,
  type = "auto", ...)
}
\arguments{
\item{x}{phylo object}

\item{id}{vector of ids or names}

\item{...}{Arguments sent to \code{subtree.phylo}}

\item{type}{id type: ['name', 'id', 'auto']}

\item{collapse}{logical collapse edges around nodes with a single descendent}

\item{descend}{logical include all descendents of each id}
}
\value{
phylo object
}
\description{
Select a subset of nodes from a tree
}
\examples{
data(atree)
subtree(atree, c('n15', 'n19'))
subtree(atree, c('t7', 't4', 't1'))
}
