% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{subtree}
\alias{subtree}
\title{Get a subtree rooted at a given node}
\usage{
subtree(tree, id)
}
\arguments{
\item{tree}{phylo object}

\item{id}{vector of ids or names}
}
\value{
phylo object
}
\description{
Get a subtree rooted at a given node
}
