% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{take_first}
\alias{take_first}
\title{Select representatives for a strata}
\usage{
take_first(x)
}
\arguments{
\item{x}{data.tree containing one stratum, especially in the NCBI common
tree case, this may contain many children.}
}
\value{
data.tree where all but the first child of each immediate child of
the stratum root is culled.
}
\description{
Select representatives for a strata
}
