% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{tree_names}
\alias{tree_names}
\title{Get the names of all leafs and nodes in a tree}
\usage{
tree_names(tree)
}
\arguments{
\item{tree}{phylo object}
}
\description{
Get the names of all leafs and nodes in a tree
}
