% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{tree_size}
\alias{tree_size}
\alias{tree_size.Strata}
\title{Get the number of species and ancestors in the tree}
\usage{
tree_size(x)

\method{tree_size}{Strata}(x)
}
\arguments{
\item{x}{phylo or Strata object}
}
\value{
integer
}
\description{
Get the number of species and ancestors in the tree
}
