% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_cousin_genomes}
\alias{uniprot_cousin_genomes}
\title{Retrive the sequences from the uniprot cousins}
\usage{
uniprot_cousin_genomes(cousins, dir = "strata", prefix = "ps_",
  dryrun = FALSE, ...)
}
\arguments{
\item{cousins}{The output of \code{uniprot_cousins}}

\item{dir}{Directory in which to write sequences}

\item{prefix}{The phylostratum-level prefix}

\item{dryrun}{If TRUE, do not download files or create directories}

\item{...}{Additional arguments passed to \code{uniprot_retrieve_genomes}}
}
\value{
Nothing, this function is run for its effects
}
\description{
cousin_sets A named list of vectors of species taxon ids. Each vector
in the list bears the name of an uncle, all at the same level in the tree.
There may be multiple uncles in multifurcating nodes of the tree (this is
very common in the NCBI common tree).
}
\examples{
\dontrun{
uniprot_cousins(3702) \%>\%
  lapply(take_first) \%>\%
  uniprot_cousin_genomes

cfilter <- make_do_if_over(3, take_first)
uniprot_cousins(3702) \%>\%
  lapply(cfilter) \%>\%
  uniprot_cousin_genomes
}
}
