% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_cousins}
\alias{uniprot_cousins}
\title{Given a focal taxid, find the uniprot descendents of a taxon's uncles}
\usage{
uniprot_cousins(taxid, ...)
}
\arguments{
\item{taxid}{The focal species NCBI taxon id}

\item{...}{Additional arguments sent to \code{uniprot_downstream_ids}}
}
\value{
list of lists of id vectors
}
\description{
Given a focal taxid, find the uniprot descendents of a taxon's uncles
}
