% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_cousins}
\alias{uniprot_cousins}
\title{Given a focal taxid, build a tree of}
\usage{
uniprot_cousins(taxid)
}
\arguments{
\item{taxid}{The focal species NCBI taxon id}
}
\value{
data.tree with strata from the NCBI lineage. Each stratum has one or
more representative clades, which are the immediate, outgroup children of
the stratum's mose recent common ancestor. For each of the 'aunts', all
descendent species that are represented in UniProt are added included. Note
this is a flat tree, the topology between aunt and descendent is lost.
}
\description{
Given a focal taxid, build a tree of
}
