% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_downstream_ids}
\alias{uniprot_downstream_ids}
\title{Get the uniprot ids downstream of a node}
\usage{
uniprot_downstream_ids(taxid, reference_only = FALSE)
}
\description{
Get the uniprot ids downstream of a node
}
