% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_fill_strata}
\alias{uniprot_fill_strata}
\title{Download sequence data for each species in a UniProt-based strata}
\usage{
uniprot_fill_strata(strata, ...)
}
\arguments{
\item{strata}{List of lists of taxon IDs}

\item{A}{named list of filename vectors}
}
\description{
Download sequence data for each species in a UniProt-based strata
}
