% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_fill_strata}
\alias{uniprot_fill_strata}
\title{Download sequence data for each species in a UniProt-based strata}
\usage{
uniprot_fill_strata(strata, ...)
}
\arguments{
\item{strata}{data.tree where leaves are named by NCBI taxon ID. These IDs
are expected to be in UniProt.}

\item{...}{Additional arguments for \code{uniprot_retrieve_proteome}}
}
\value{
A data.tree with the additional field 'protein_file'
}
\description{
Download sequence data for each species in a UniProt-based strata
}
