% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_retrieve_proteome}
\alias{uniprot_retrieve_proteome}
\title{Retrieve UniProt proteome}
\usage{
uniprot_retrieve_proteome(taxid, dir = "uniprot-seqs", verbose = TRUE)
}
\arguments{
\item{taxid}{An NCBI taxonomy id}

\item{dir}{Directory in which to write all FASTA files}

\item{verbose}{Be super chatty}
}
\description{
The proteome is written to a file with the name '<taxid>.faa', for example,
Arabidopsis thaliana, which has the id 3702, will be written the '3702.faa'.
}
\examples{
\dontrun{
# uniprot_retrieve_proteome(3702)
}
}
