% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot-access.R
\name{uniprot_sample_prokaryotes}
\alias{uniprot_sample_prokaryotes}
\title{Randomly sample prokaryotic representatives}
\usage{
uniprot_sample_prokaryotes(downto = "class")
}
\arguments{
\item{downto}{the lowest phylogenetic rank that shoul be sampled}
}
\value{
phylo object containing the prokaryptic sample tree
}
\description{
Randomly sample prokaryotic representatives
}
